requirements.txt
cohere==5.3.3
google-generativeai==0.5.2
llmsherpa==0.1.4
pangea-sdk==3.8.0b2
streamlit==1.33.0